xml: reorg
